streamlit_app/requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
blinker==1.7.0
boto3==1.28.78
botocore==1.31.78
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
geographiclib==2.0
geonamescache==2.0.0
geopy==2.4.0
gitdb==4.0.11
GitPython==3.1.41
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
idna==3.7
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.7
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.3.0
proto-plus==1.22.3
protobuf==4.25.2
pyarrow==14.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
PyPrind==2.11.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.1
requests==2.32.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
s3transfer==0.7.0
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.36.2
streamlit==1.37.0
streamlit-javascript==0.1.5
streamlit-option-menu==0.3.6
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.3
wheel==0.41.2
zipp==3.19.1
zope.interface==6.1
